Fix bug in hangman.py
The script is supposed to randomly select a word (the variable for this is ran_word) and convert that same word into a list of letters (the variable is word). However, the script was randomly selecting a word and then turning *another* randomly selected word into a list of letters. The error was fixed by not calling the choose_word() function in the word_to_list() function and by passing in ran_word as a parameter to word_to_list().