Add the you_won() function to hangman.py
The you_won() function was created to avoid repetition. A user can win the game after choosing a letter that appears only once in the word or after choosing a letter that appears twice in the word; each scenario is handled by different if statements. That means that an if statement handling whether or not the user won was reapeated once in the play_game() function. This repetition has now been removed.